--- src/main/java/controller/dto/WinningResult.java
@@ -1,6 +1,0 @@
-package controller.dto;
-
-import player.Name;
-
-public record WinningResult(Name name, boolean isWinner) {
-}
