--- src/main/java/dealer/dto/DealerWinningResult.java
@@ -1,4 +1,0 @@
-package dealer.dto;
-
-public record DealerWinningResult(int winningCount, int failCount) {
-}
