+++ src/main/java/participant/player/Name.java
@@ -1,4 +1,4 @@
-package player;
+package participant.player;
 
 public class Name {
