--- src/main/java/player/dto/CardsStatus.java
@@ -1,11 +1,0 @@
-package player.dto;
-
-import card.Cards;
-import player.Name;
-
-public record CardsStatus(Name name, Cards cards) {
-
-    public int getCardsScore() {
-        return cards.countMaxScore();
-    }
-}
