+++ src/test/java/participant/player/NameTest.java
@@ -1,4 +1,4 @@
-package player;
+package participant.player;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
